Added comments to files
The Modeling of the business logic here does not make it apparent what the intention or purpose of this application is. logic has been spread out throughout the application making it difficult to have consistency. business objects do not own their behaviors. Services are performing non service Functions and Controllers are not taking care of providing actionable intel. One thing that stands out is the separation from business models and DTOs propagated to the consumer.

Some concerns are in the wrong place and need to be moved around. I can not see the need for the Application Project. Services must not concern themselves with the CRUD operations. Special consideration must be paid to the Repository pattern + UOW seeing we are using entity framework.

Bounded contexts are not apparent (DDD Eric Evans)

For brevity, I will leave this for an in person discussion